requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
arrow==0.15.2
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
awscli==1.16.284
binaryornot==0.4.4
black==19.10b0
boto3==1.10.20
botocore==1.13.20
certifi==2019.9.11
cffi==1.13.2
Click==7.0
colorama==0.4.1
cookiecutter==1.6.0
cryptography==2.8
docutils==0.15.2
Flask==1.1.1
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-Testing==0.7.1
importlib-metadata==1.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jinja2-time==0.2.0
jmespath==0.9.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
pathspec==0.7.0
poyo==0.5.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.3
pynamodb==4.2.0
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
regex==2020.1.8
rsa==3.4.2
s3transfer==0.2.1
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.11.2
zipp==0.6.0
